Fix stage names in HttpAPI
Terraform doesn't expose the invokeUrl or executionArn on
a Stage - only a Deployment - even though the Stage is the
only thing that actually has an endpoint associated with it.

We have to use a hack of providing an empty string to the
Deployment's "stageName" property, and then adding the
stage name back into the invokeUrl and executionArn.