Simplify default integration example.
Adds multiple entrypoints to the integration sample
so that we can highlight a variety examples.  Makes
the simplest the default.

Once we've fleshed out enough in this example, we'll
likely move pieces of this into other libraries.  But for
now, it's useful to experiment with different capabilities
in a sample.