Support binary content in HttpAPI
Fixes #21.  Enables binary data to be served from `routeStatic` as well as passing through
raw binary data on both requests and response body payloads for function handlers.

API Gateway supports this only via transformation to the binary input to base64 and back. As
a result, we tell API Gateway to treat all inputs as binary, and then base64 decode them into the
Buffer we pass to users as the request body.  Similarly, we collect the response buffer, then base64
encode it to hand back to API Gateway, which will transform back into binary.

It's a little crazy that we have to do all this encoding/decoding, but API Gateway does not appear
to support binary content  lambda integrations without this.