Support for attaching custom domains to HttpAPI
This adds support for an `attachDomain` function on HttpAPI.

This will allow a custom domain, along with user-provided SSL/TLS certs, to be set up as part of the API deployment.

`attachDomain` returns a domain name (in current implementation it is a CloudFront distribution) that the customer must point their domain at using an A record.

This requires that customers provide their SSL/TLS certificates.  We can support doing so via config instead of in source code, but there's a question about whether this is acceptable in general. (It's similar to what most related systems do, albeit throug config web UI).