[WIP] Introduce PulumiResources abstraction
Adds a Go package for inferring the Pulumi Framework resources from a Lumi checkpoint file.

The structure of PulumiResources intends to be the form that we can use to communicate to users about the resources in a Pulumi Framework application.

We expect to expose functions for extracting metrics (and logs) for each of these resources on this API.

Future implementations of the Pulumi Framework targeting other environments (Azure, Kubernetes, etc.) will be expected to be able to expose the same interface, but will implement the deserialization of a checkpoint file into this interface differently (and in the future, the implementation of metrics and logs collection).

Initially, we use this just to replace the runtime testing discovery of RestAPI endpoints.