Add seperate API definition files for the pulumi-framework nodejs API. (#52)
Add seperate API definition files for the pulumi-framework nodejs API.

This code lives in the "api" directory and should become the canonical place where we define our api.  Think of it like the idl for pulumi-framework.

There are also two implementations of this code.  One in 'aws' and one in 'local'.  As the names imply, these are the implementations that can run in aws using amazon cloud services, and locally, using simple node package to provide services like HttpAPI.  In the future we can add implementations like 'azure' and 'gcp' as appropriate.

Also in 'api' is a micro module loader.  When the api module actually is loaded in node it will dynamically determine which environment it is running in, and will then reexport the right implementation for that environment.  it is the responsibility of the deployment system to appropriate place the right information in the environment, and include the implementation package for that enviornment so that node can find it.

For example, in aws the deployment system would place pulumi/pulumi-framework-aws in the node_modules directory.  pulumi/pulumi-framework would tehn dynamically discover it was in aws and would import pulumi-framework-aws and then reexport it.