Directly implement API interfaces on the implementation types. (#59)
Directly implement API interfaces on the implementation types.

This improves the development experience by moving typescript validation
and error reporting closer to the source of the problem.  Without this,
api conformance happens only in index.ts and can result in a large and
difficult to read message.

This does mean conformance error messages are effectively duplicated.  But
that shouldn't be too bad.