Fix `make` to build everything but not run tests
This logically reversts my previous change, moving us back to
$(MAKECMDGOALS). This was done so that `make` (which builds the
`default` target) does not build the `all` target in the
subprojects. This lets a subproject have different behavior between
make and make all (which the aws project uses to not build the
examples when only the default target is built).