First step on containers
This implements the first small step on supporting containers.

This first example exposes a @pulumi/cloud notion of Service - a container(s) which will run permanently for the lifetime of the program, and which may expose some ports whcih can be accessed from other compute (aka functions).

To use Service, you must configure the `cloud-aws:config:ecsClusterARN` variable to point to a valid ECS cluster in the region you are deploying into.  This is effectively treating that as part of the infrastructure that an execution environment must make available.

The Service also exposes a runtime API to provide it's host/port pairs for each exposed port on the container(s).  This is currently handled by querying the ECS data plance for info about the single instance of the service.  We expect to replace this with an NLB to ensure this service discovery is more efficient (and more robust/managed).

Notably, the example shows how a lambda can make requests against a running service.

All APIs defined so far are provisional - we'll flesh these out and simplify+enrichen as needed based on the next round of examples.

We have lots more to do for containers - including:
* [ ] Move to NLB and support scaling services
* [ ] Solve for two services defined by external images discovering one another
* [ ] Support for doing Docker builds as part of provisioning (`build` vs. `image`)
* [ ] Support for closure serialization into Docker image transparently
* [ ] Support for notion of Task which is long running job that can be spawned dynamically and will finish
* [ ] Potentially allow using Task as a function