Support for containers defined by JavaScript lambdas
Allows a `function` to be specified in place of an `image` in a container specification.

Uses closure serialization to turn this into JavaScript which can be loaded into a fixed Docker runner image.

Includes the source for the `javascriptrunner` Docker image which is published as `lukehoban/javascriptrunner`.  We will need to publish this as an official Pulumi image.

Also doesn't yet pass the full local folder zip into the container - which is necessary for correct closure serialization semantics.  But many/most simple examples will already work.