Refactor to use ComponentResource base class
This makes some fairly large changes to the way the cloud framework
resources are structured, so that we can leverage the ComponentResource
base class for auto-parenting and the like.

The largest change is the introduction of an HttpDeployment resource.
The HttpEndpoint class is really just a factory for deployments.  As
part of this, we now defer creation of all actual resources until
publish is called.  This makes it safe to call publish multiple times
if desired, and ensures that we can parent everything nicely.

The rest of the changes are for the most part fairly straightforward.
In general, all resource initialization occurs in the callback passed
to the ComponentResource base class and any subsequent assignments,
like creation of "inside" APIs, typically occurs afterwards.