Automatically provision ECS clusters
Previously, we threw an error if you tried to create a Container or
a Task and didn't manually configure an ECS cluster ARN by provisioning
one ahead of time and then setting cloud-aws:config:ecsClusterARN.

This change introduces automatic creation of a canonical ECS cluster
on-demand.  If you choose to manually configure one in the same manner
as before, we will use it, but this gives you a reasonable default.

There are a few knobs you can decide to use if you want:

* cloud-aws:config:ecsAutoClusterInstanceType overrides the instance
  size from its default of "t2.micro".

* cloud-aws:config:ecsAutoClusterDesiredCapacity overrides the
  desired capacity count from its default of 2.

* cloud-aws:config:ecsAutoClusterMinSize overrides the minimum
  cluster size for the ASG, which uses the desired capcity by default.

* cloud-aws:config:ecsAutoClusterMaxSize overrides the maximum
  cluster size for the ASG, which uses the desired capcity by default.

* cloud-aws:config:ecsAutoClusterDisable disables the creation of
  auto-clusters, which is reasonable if you want to avoid mistakes.

It's not clear what we'll do with this long-term, but it certainly
makes our demo and simple container scenarios sing!