Support for container build
Adds support for the `build` mode of deploying container-based compute (Service or Task).

The build is done using the locally available `docker` engine via the Node.js Docker API (dockerode).

The build process is run on any update, but is very quick in general due to Docker caching.  The SHA digest for the image is attached to the environment variables of the TaskDefinition which causes that AWS resource to be updated IFF the image actually changes.  This triggers the service to flush tasks running with the old image and update to the new one.

There is a lot of work remaining to clean things up - including ensuring failure modes all report useful errors and that status is communicated cleanly.