Add unique suffixes to LogGroups
The LogGroup we create for cloud framework functions was given the
name of the function, and the LogGroup we create for container services
was also given the name of the service.  Unfortunately, this means if
you create a function and a service with the same name, you get an obscure
"names aren't unique" error message due to the LogGroups.  We can avoid
this (for now) by adding suffixes to the names.