Support for auto-creating Clusters
Adds the ability to create an ECS cluster automatically if one is not provided.

Builds upon the Network abstraction to provision the Cluster in either a Public or Private subnet depending on Network configuration.

When the cluster is running in a private subnet, `Services` will expose internal-only load balancers that are visible to Lambda and Containers, but not directly to the internet.  This allows HttpEndpoint's to be used as the only directly internet-visible endpoints on the VPC.

EFS support for `Volumes` is not yet enabled.