Build containers, even during 1st-time previews (#163)
The prior code wouldn't build containers until we had a repository URL
in hand, which meant until an actual deployment had occurred, we would
skip several arguably important steps.  (E.g., you wouldn't find Docker
build failures until attempting to deploy.)

This change builds the image even the first time around, using our
automatic name for it.  It then explicitly tags the result, when it has
the repository URL, and pushes that.

We also now permit you to specify an image *and* a build directory and
we will use that image rather than auto- computing it.