Fix circular resource dependency
The current service code can lazily initialize an ECR registry
from within the component initialization for the service that
depends on said registry.  Because we serialize outbound RPCs,
this has the effect of the service blocking on properties of
the registry ... which won't actually be resolved until the
registry has been created ... which won't happen because that
is dependent on the creation of the service.  Deadlock!

The failure mode is pretty terrible here, and I will go ahead
and file an issue with details momentarily.

In the meantime, we can just make sure we don't end up with
such reentrant circular dependencies.  This looks a little bit
more like the old code, but we still get the two desired effects
of the recent changes: (1) we can share repositories for images
with the same name, and (2) we will still build -- without
publishing -- during previews.