Ensure ECR Repositories are created ahead of TaskDefinitions
This finally fixes the racy promise issues we've been having in
the cloud repo since some of the recent container/service changes.
The root cause here turns out to be the following sequence of events:

1) We start to create a cloud.Service.

2) Being a cloud component, this calls runtime.initResource.  It then
   proceeds to invoke the component init function, without awaiting it.
   (It cannot await it, since this invocation occurs in a ctor.)

3) This init function is what allocates the constituent parts, like
   aws.ecr.Repositories and aws.ecs.TaskDefinitions.

4) If we are able to new up the aws.ecr.Repository *first*, then it
   will get slotted ahead of the dependent resources that are allocated
   later on (and which depend upon it).

5) If we do not, however, due to (say) an asynchronous interleaving,
   then the aws.ecs.TaskDefinition which requires the registry ID,
   repository URL, etc., will end up slotted at the head, at which
   point it blocks on the repository's properties and we deadlock/leak.

This change delicately sequences things so we know the aws.ecr.Repository
will win the race.  Obviously this is supremely awful.  The work we
have planned for M9 around blocking and components should rectify things.

Committing without a PR since the branch is on the floor and tests
take forever these days.  Hopefully it unsticks things.