Use repository name in TaskDefinition (#177)
We need to use the repository name in the task definition so that we can access the image from ECR.

Also fixes issue with intermittent Travis failures.  It appears that occasionally the `exit` event was allowed to fire before the subprocesses `stdout` `end` event, leading to not returning the correct stdout contents.