Expose a getLoadBalancer on AWS Services
This change exposes a getLoadBalancer function so that callers can
use it when wiring up other AWS infrastructure.  For instance, we
need this in the customer stack to point the CDN at the web stack.

To access it, you can do, for example

    import * as cloud from "@pulumi/cloud";
    import * as awscloud from "@pulumi/cloud-aws";
    let svc = new cloud.Service(...);
    let lb = (svc as awscloud).getLoadBalancer();

I've tried to do this in the most elegant way possible, with the
caveat that this of course needs to be revisited as part of
pulumi/pulumi-cloud#145 alongside the other similar things.