Permit Application-level load balancing, and SSL
This change permits Application-level load balancing, which
permits things like SSL termination, HTTP/2, and WebSockets.

A cloud.Service may now declare the protocol for any one of
its ports: "tcp", "http", or "https" are supported.  In the
case of "tcp", we keep the current Network-level load
balancing scheme, while in the case of "http" or "https", we
chosen Application-level elastic load balancer.

We continue caching these as before.

Because a certificate is required for the SSL termination
at the Application load balancer, we have a new configuration
variable that may be set, acmCertificateARN, which will be
automatically fed into load balancers as appropriate.  There
is currently just one global one per stack.

I'm also changing the default number of AZs to 2, rather than
1, although I understand there may be Network load balancer
troubles introduced by this that we'll need to address before
merging in the rest of these changes.