Eliminate stdout spew for "docker version"
The code recently added to capture the Docker version didn't ask to
redirect the stdout, so it would appear in the actual output of any
Pulumi program that used a service.  I don't think this was intentional,
and in fact, the version doesn't appear to be used for anything yet.
So, passing true to capture the version, although based on what got
printed to the console, it's going to require some reworking...