Reduce default IAM privelege
Reduces the default EC2 Instance Role policy to the standard AmazonEC2ContainerServiceforEC2Role policy, and adds support for overriding it in `Cluster` and in the `@pulumi/cloud` config.

Also reduce default compute permissions to just `AWSLambdaFullAccess` as the previous additional priveleges are not required by our frameworks.

Also exposes a config variable to suppress creation of EFS as part of an auto cluster.  Fixes #175.

Related to pulumi/pulumi-ppc#89.