Reduce default IAM privelege (#261)
Reduces the default EC2 Instance Role policy to the standard AmazonEC2ContainerServiceforEC2Role policy + readonly EC2 access, and adds support for overriding it in `Cluster` and in the `@pulumi/cloud` config.

Also exposes a config variable to suppress creation of EFS as part of an auto cluster.  Fixes #175.

Related to pulumi/pulumi-ppc#89.