Allow configuring AWS IAM role programmatically
We previously allowed AWS-specific configuration only through Pulumi config.  This is quite limiting, and prevents cases where AWS resources need to be created as part of a Pulumi program and then used as art of initializing the `@pulumi/cloud` AWS implementation.

With this change, we introduce a first AWS-specific API `setComputeIAMRolePolicies`. This can be accessed explicitly via the `@pulumi/cloud-aws` package.