Increase service deregistration delay (#304)
We have seen cases where the 30s delay between LB deregistration and stopping the container is not enough for the new copy of the service to be active, leading to downtime during updates.  This is not a full solution to pulumi/pulumi-ppc#139, but should help avoid serious issues until we can reactor the PPC.  It also provides a less surprising default for the cloud framework.

Also allow promises for a few additional container inputs, and expose a few additional properties on the AWS variant of the cloud Service.