Expose additional AWS resources in @pulumi/cloud-aws
Now that @pulumi/cloud-aws is a separate API from @pulumi/cloud, we can safely expose AWS implementation details in the public API surface of the former.  This has proved important for interoperating with other AWS capabilities in AWS-specific environments.

Here, we expose additional underlying implementation details for Topic, Timer, Table and HttpEndpoint.

We also turn on detailed metrics for API Gateway so that we can colelct per-route metrics in CloudWatch.

Fixes #316.