Support for `proxy` in HTTP endpoints (#328)
Support in HTTP Endpoint for proxying to either a URL or a cloud.Endpoint.

This is part of addressing pulumi/pulumi-ppc#164, and generally enables a route on an HTTP Endpoint to redirect through to another HTTP endpoint - either a URL on the public internet, or an endpoint inside the private network of an @pulumi/cloud stack.

The API for exposing this is aligned with how we currently support static routes. If/when we address #4, #188, and the related set of issues around a redesign of HttpEndpoint to either align more closely with Express or with API Gateway - we may choose to adopt a different user-facing API for both static and proxy routing. But for now - these both map to the ability to route to sources which do not need to run code - static files (in S3), or proxied content (in an NLB in a VPC, or at a fixed internet address). For support on other cloud platforms without a managed API Gateway, we may need to run custom runtime code to do the proxying, which we expect will be fine if necessary (similar to the static route support, which may need custom code to pull content from storage).

Fixes #287.