Support for ACM certificates on HttpEndpoint (#343)
We allow the AWS implementation of HttpEndpoint#attachCustomDomain to accept an ACM cert in place of raw certificate material.  This greatly eases the ability for AWS users to take advantage of HttpEndpoint with custom domains.

This has come up as a requirement for several places we want to use HttpEndpoint for our own web properties and APIs, but need to be able to serve at a specific custom domain.

Fixes #268.