Add parallelism to unit tests
The runtime validation was being done sequentially.  This is particularly problematic, as we have a 30s timeout on the HTTP request that does this validation.

Until/unless we move to an asynchronous model where tests are queued and executed out of band of an HTTP request, we should at least make sure to run tests in parallel so that we only have to worry about the long pole of an individual test taking more than 30s.

This change reduces cold invocation from >30s to <10s, and warm invocation to <4s.