Ensure Deployment is recreated on all changes to API body
Previously, we were using the name of the Deployment as the way we
would trigger replacements of the Deployment when the body of the
API changed.  However, this required that we be able to compute the
body hash promptly, prior to any deployment progress - which meant
we left some information (any Computeds) out of the hash.

This leads to problems when the only change to the API is to one of
those Computed values.  In that case, we would not end up redeploying
the API, leading to a bad state.

Instead, we now use the `variables` property which is ForceNew to
trigger replacement when the hash changes.  We can pass the value
for this property as a Computed, which means we can compute the hash
based on the fully realized body, includin the final values of
Comuteds.

We still use a blank stageName on the Deployment so that we do not
actually create any unnecessary user-visible Stages beyond the one
stable stage where the API is published.