Support custom Docker build args
This change adds support for custom Docker build args.  This is "inspired
by" Docker Compose's approach of permitting you to either just specify a
string -- which is interpreted as the context -- or a bag of options.  The
options specified may contain the context, a custom Dockerfile location
(just a shortcut for the common `-f <path>` args), or a bag of args, which
may include anything that the Docker build command itself supports.