Disable verbose setting on containers test
By default, we only emit output if a test fails.  This makes it easier
to follow along with what's happening; on success, we are silent, and
on failure, we will see precisely and only what went wrong.  We seem to
have flipped verbose to true to help deal with some Travis issues, however
I would rather see us fix the underlying issue, and/or break tests into
smaller units, than to make debugging failures in the normal case harder.
We also now have the S3 buckets to fall back on if we need to.