Switch to consuming NPM packages
This change adds real dependencies to our NPM packages in the primary
packages, @pulumi/cloud, @pulumi/cloud-aws, and @pulumi/cloud-mock, in
this repo.  Note that any inter-dependencies amongst these lock to the
same version, since we are publishing from the same repo.  The single
explicit dependency on a lower layer package, @pulumi/cloud-aws consuming
@pulumi/aws, does so with a specific package version (rather than using
the prior Gopkg.* method for locking to a commit hash).  All examples
and tests are also updated to have real package dependencies.