Specialize @pulumi/cloud-aws's package.json
This change addresses some issues in the way @pulumi/cloud-aws depends
upon other Pulumi SDK packages.  Namely, the way it was written before
led to strange /opt/pulumi/node_modules/ layouts and improperly
supported inner loop symlinking.  (We *hope* much of this goes away as
soon as the core pulumi package is fully on an NPM plan...)

Specifically, this change:

* Make pulumi and @pulumi/cloud peer dependencies.

* Make @pulumi/aws a peer dependency within @pulumi/cloud-aws by default.

* Add a package.json.publish file, which lists @pulumi/aws as a real
  dependency at the appropriate commit, for use when publishing to NPM.

* Add back the @pulumi/aws entry in the Gopkg.toml so that we can use
  the same workflow as developers do to symlink using @pulumi/cloud-aws's
  default package.json file.