Publish with real dependencies
This change adopts our new workflow for publishing the cloud packages
to NPM that can be used entirely with npm install, and notably no symlinking.
Because all of our tests still use symlinking *within* the SDK, we still use
peerDependencies in most places.  Upon publishing, we will promote the
peerDependencies of @pulumi/pulumi and @pulumi/aws to real ones.