Require provider config and improve error message when provider not installed (#377)
To be more explicit about how `@pulumi/cloud` picks up it's dependency on AWS, we now make this behaviour non-default - requiring users to specify `cloud:config:provider` explicitly. Currently only `aws` is supported, but this ensures that the connection is understood.

Now that users must specify Node.js package dependencies explicitly instead of using an SDK, it is also more common to forget to include the dependency on the target cloud implementation of @pulumi/cloud - typically @pulumi/cloud-aws.

Here we improve the error message in this case - explicitly noting the module that needs to be installed.

Fixes #394.