Minimal changes to support correct dependency tracking for Network
This builds on top of the in progress PR but simplifies the changes for now.

Instead of making every output property depend on the whole component being created, we just address the specifc case where a resource is not at all usable until another resource has been created - the Subnet and RouteTableAssociation.

We also can get rid of all the preivous manul "dependsOn" logic, because the dependencies on the internetGateway and natGateways are carried along with the Output<T> for the subnet now.