Explicitly create execution role instead of assuming it already exists
AWS auto-creates a role for this purpose on ECS console first-run experience and a few other conditions - but we can't rely on that.

We also can's easily use the existing role if it exists but fall back to creating one.

So instead, we must create a new role for every stack.