Undo fix for #402
We attempted to fix #402 by explicitly creating the LogGroup for log collector lambdas, so that it would be managed by Pulumi and could be deleted when the program is destroyed.

However, for existing stacks, the Log Group will already have been created, and attempting to re-create it leads to an error.

For now, we'll just undo the fix for #402 and reopen that issue to explore alternative fixes.