Support Fargate (#411)
Adds a `cloud-aws:useFargate` flag which causes container compute to run in Fargate.

Also uses the default VPC as the target network by default (when neither `cloud-aws:externalVpcId` nor `cloud-aws:usePrivateNetwork` are defined).

This significantly simplifies the number of resources needed for simple container examples - by avoiding both EC2 and networking resources. Users can switch to private networks with usePrivateNetwork or to their own EC2 instances with ecsAutoCluster. We continue to need to clean up this config matrix, which is tracked in #134.

Fixes #410.