Do not push to registry during preview
When we improved our use of "IsStable" to avoid noise during preview, it appears that it caused our current logic for deciding when to push to a Docker registry to want to push new bits during preview - because the url of the target registry is "Stable".

Instead, we explicitly check for whether we are doing a preview, and skip the push in that case.

Longer term, it may make sense to treat the docker push as a dynamic resource whose update operation does the push, so that this falls out more naturally in the execution model.

Fixes #445.