Fix non-Fargate support for `cloud.Service`
Update to use the latest ECS AMI with AutoCluster.  Per discussion with AWS support, it appears they broke something such that `awsvpc` mode doesn't work on older AMIs (despite the documentation saying that is does work with the combination we were using).

Switch to `ip` base load balancer targeting even in non-Fargate mode which is required now that we are using `awsvpc` networking mode.

Finally, remove the `iamRole` passed to ECS Service - as ECS with `awspvc` networking requires that the service-linked-role be used, which does not allow also specifying a custom IAM role.

Adds additional test coverage to run the `containers` test both on Fargate and using AutoCluster.

Also changes `ecsAutoClusterUseEFS` default to be false since this is expensive to create and not used by any current users of AutoCluster.