Update dependencies to latest pulumi{-aws}
We recently took a dependency on a breaking change in @pulumi/pulumi, but did not update our dependency.  This was okay in Travis because we still use `link` there - but fails when this package is used via NPM.

This change updates our dependencies for `api` and `aws` packages, and moves all examples to a more generic version specification.

Note that the `0.11.1-dev` versions mentioned here are actually *newer* than the `0.11.2` builds and contain the fixes, and so this change to try to require these newer builds may not actually work.