Allow paging in Table.scan.
Add an overload of Table.scan that accepts a callback that will be
invoked as each page of results is fetched. If this callback returns
`false`, no more pages will be fetched.

These changes also fix a bug wherein the AWS implementation of
Table.scan would not fetch all results if they had been paginated.