Remove LogCollector and onError (#474)
These were added early on as part of a much more PaaS-like view of `@pulumi/cloud`, but in practice have not added value for users, and add complexity to all deployments using `@pulumi/cloud`.

We may choose to bring one or both of these concepts back to `@pulumi/cloud` - most likely as opt-in, in the future.  But for now, we should remove them and keep things simpler by default.

Fixes #402.
Fixes #42.