Rename HttpEndpoint to API (#497)
Use a simpler name that is more aligned with the common naming in the space.

Note that a type/value with the name HttpEndpoint are still exported, as aliases for API, but marked as `@deprecated`.  This will ease adoption of the new name.  We will plan to remove these in a future major release.

Fixes #476.