Add a simpler way to create container services (#535)
This change adds a simpler way to construct container services. The
cloud.Service abstraction permits multi-container services, however that
is a more advanced scenario than the simple case of just wanting to
stand up a single container. This unfortunately means you have to give
the container a name which is almost always repetetive with respect to
all the other names floating around.

For example, this

    let nginx = new cloud.Service("nginx", {
        containers: {
            nginx: {
                image: "nginx",
            },
        },
    });

simply becomes

    let nginx = new cloud.Service("nginx", {
        image: "nginx",
    });

The approach used simply lets the Container information to come in with
the ServiceArguments. If that is present, we simply create a container
map out of it, using "default" as the name. We don't support mixing.