Remove performance tests
We are not monitoring the results of these tests and they don't really
make sense in our new thinking where these abstractions are very
lightweight wrappers over the underyling cloud infrastrcture. Because
of that, these ends up really testing the performance of the
underyling cloud resources (e.g. DynamoDB) instead of something Pulumi
is able to influence.