Promote @pulumi/cloud peerDependency to dependency when publishing
We must have the reference to @pulumi/cloud be a peerDependency in our
checked in package.json file, because we want to depend on the version
we have just built (which is not on NPM).

But, we don't want it to be a peerDependency when we go to actually
publish the package. So we use the `promote.js` technology to update
the peerDependency to a real dependency at publish time.