Work around breaking change in invoke
The recent change to not allow Output<T>'s to be passed to data
sources broke this code. The quickest way to unblock ourselves is to
cheat and get at the underlying promise, await it and then pass the
data to the data source. This is logically what the old code would
have done anyway.

Issue #544 tracks doing better here.