Use On Demand for AWS Table (#679)
The AWS Tabe implementation used DynamoDB with an opauqe and not-configurable fixed provisioning.  This has been probematic for several reasons:
1. It does not scale with load
2. It incurs a fixed cost even when not used
3. It would require AWS-specific provisioning hooks to expose control

With the recent release of On Demand billing mode for DynamoDB, we can now avoid all three problems by setting Table to On Demand provisioning which will scale with demand (including to zero).

We may in the future allow AWS-specific options to override this, but this makes much more sense as the default for this library.

Note that the change from Provisioned to On Demand does not require replacement of the Table, it can be updated in place.  The update itself does take ~5minutes to complete, but it otherwise non-disruptive to use of the Table.

Fixes #31.